+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- Dataminer module included.
+- Created by and Updated by on labels.
+- Labels can be archived and un-archived (active/inactive).
 ### Changed
+- Status logging for create, link and delete of labels.
+- List of variables for a label can be configured via the variable set and shared_config.
 ### Fixed
 
 ## [0.3.0] - 2019-01-04